requirements.txt
pypng==0.0.17
Pillow==2.5.0
